task_manager_api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymongo==4.6.2
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.31.0
